Add side panel

+++ src/com/vl/marketing/view/DashboardController.java
@@ -156,16 +156,17 @@ public class DashboardController {
 //			table.edit(tp.getRow(),  tp.getTableColumn());
 //		});
 
-		// TODO Replace isShiftDown() with proper login system
 		table.setRowFactory( tv -> {
 			TableRow<Authorization> row = new TableRow<>();
 			row.setOnMouseClicked(event -> {
 				Authorization rowData = row.getItem();
-				if (!event.isAltDown() && !event.isShiftDown() && event.getClickCount() == 2 && (!row.isEmpty()) ) {
+				if (!event.isAltDown() && event.getClickCount() == 2 && (!row.isEmpty()) ) {
 					setClickedStatus(rowData.getStatus());
-					main.showNewAuthorization(this, rowData);
-				} else if(event.isShiftDown() && event.getClickCount() == 2 && (!row.isEmpty())) {
-					main.showApproval(this, rowData);
+					if(user.getPrivileges().get("submit")) 
+						main.showNewAuthorization(this, rowData);
+					else
+						main.showApproval(this, rowData);
+				// TODO Replace pdf generation with a button
 				} else if (event.isAltDown() && event.getClickCount() == 2 && (!row.isEmpty())) {
 					PDFGenerator pdfGen = new PDFGenerator();
 					try {
@@ -260,19 +261,23 @@ public class DashboardController {
 		ObservableList<Authorization> toDelete = FXCollections.observableArrayList();
 		Authorization auth;
 		
-		for(int i = 0; i < authorizations.size(); i++){
-			table.getSelectionModel().select(i);
-			auth = table.getSelectionModel().getSelectedItem();
-			
-			// Only pending and rejected authorizations will be rejected
-			if(auth.getChecked() == true && (auth.getStatus().equals("pending") || auth.getStatus().equals("rejected"))) {
-				toDelete.add(auth);
-			};
-		}	
-		if(toDelete.size() > 0) {
-			database.removeAuthorizations(toDelete);
-			refreshTable();
-			toDelete.clear();
+		if(user.getPrivileges().get("delete")) {
+			for(int i = 0; i < authorizations.size(); i++){
+				table.getSelectionModel().select(i);
+				auth = table.getSelectionModel().getSelectedItem();
+				
+				// Only pending and rejected authorizations will be rejected
+				if(auth.getChecked() == true && (auth.getStatus().equals("pending") || auth.getStatus().equals("rejected"))) {
+					toDelete.add(auth);
+				};
+			}	
+			if(toDelete.size() > 0) {
+				database.removeAuthorizations(toDelete);
+				refreshTable();
+				toDelete.clear();
+			}
+		} else {
+			System.out.println("Cannot delete authorizations");
 		}
 	}
 	
@@ -282,18 +287,22 @@ public class DashboardController {
 		ObservableList<Authorization> toAuthorize = FXCollections.observableArrayList();
 		Authorization auth;
 		
-		for(int i = 0; i < authorizations.size(); i++){
-			table.getSelectionModel().select(i);
-			auth = table.getSelectionModel().getSelectedItem();
-			if(auth.getChecked() == true && auth.getStatus().equals("pending") || auth.getStatus().equals("rejected")) {
-				toAuthorize.add(auth);					
-			};
-		}	
-		
-		if(toAuthorize.size() > 0) {
-			database.updateStatus(toAuthorize, "approved");
-			refreshTable();
-			toAuthorize.clear();
+		if(user.getPrivileges().get("approver")) {
+			for(int i = 0; i < authorizations.size(); i++){
+				table.getSelectionModel().select(i);
+				auth = table.getSelectionModel().getSelectedItem();
+				if(auth.getChecked() == true && auth.getStatus().equals("pending") || auth.getStatus().equals("rejected")) {
+					toAuthorize.add(auth);					
+				};
+			}	
+			
+			if(toAuthorize.size() > 0) {
+				database.updateStatus(toAuthorize, "approved");
+				refreshTable();
+				toAuthorize.clear();
+			}
+		} else {
+			System.out.println("Cannot approve authorizations");
 		}
 	}
 	
@@ -341,6 +350,10 @@ public class DashboardController {
 	
 	public void setUser(User user) {
 		this.user = user;
+	}
+	
+	public User getUser() {
+		return user;
 	}
 	
 	public String getClickedStatus() {
